feat(ffi): implement form field FFI bindings for Node.js
- Add native/form.cc with 18 form widget N-API functions
- Implement pdf_widget_* functions in nanopdf-rs/src/ffi/form.rs
- Fix lifetime errors by refactoring closure chains
- Add form widget declarations to mupdf_minimal.h
- Fix fz_quad field access in stext.cc (ul/ur/ll/lr instead of x0/y0/etc)
- Fix ambiguous BigInt::New calls by explicit uint64_t casts
- Update binding.gyp and nanopdf.cc to include form module
- Add FieldFlags::has() method for flag checking
- Remove summary documents per user request

Form FFI includes:
- Widget navigation (first, next, drop)
- Widget properties (type, name, rect, value)
- Widget state (readonly, required, valid)
- Text field specific (format, maxlen, multiline)
- Checkbox/radio specific (checked state)
- Choice field specific (option count, options)
- Widget updates (appearance)

+++ nanopdf-js/native/annot.cc
@@ -23,7 +23,7 @@ Napi::BigInt CreateAnnotation(const Napi::CallbackInfo& info) {
     if (info.Length() < 3) {
         Napi::TypeError::New(env, "Expected 3 arguments: ctx, page, type")
             .ThrowAsJavaScriptException();
-        return Napi::BigInt::New(env, 0);
+        return Napi::BigInt::New(env, static_cast<uint64_t>(0));
     }
 
     bool lossless;
@@ -35,7 +35,7 @@ Napi::BigInt CreateAnnotation(const Napi::CallbackInfo& info) {
     if (annot_type < 0 || annot_type > 27) {
         Napi::TypeError::New(env, "Invalid annotation type (must be 0-27)")
             .ThrowAsJavaScriptException();
-        return Napi::BigInt::New(env, 0);
+        return Napi::BigInt::New(env, static_cast<uint64_t>(0));
     }
 
     // Call Rust FFI
@@ -480,7 +480,7 @@ Napi::BigInt CloneAnnotation(const Napi::CallbackInfo& info) {
     if (info.Length() < 2) {
         Napi::TypeError::New(env, "Expected 2 arguments: ctx, annot")
             .ThrowAsJavaScriptException();
-        return Napi::BigInt::New(env, 0);
+        return Napi::BigInt::New(env, static_cast<uint64_t>(0));
     }
 
     bool lossless;
